Emit dots on travis to avoid timeout
The current method of running a background sleep until travis is
finished is causing integration test runs to pass even when they should
be failing.

Instead, update the code to emit dots itself.
