--- tests/integration_tests/log_utils.py
@@ -1,11 +1,0 @@
-def verify_ordered_items_in_text(to_verify: list, text: str):
-    """Assert all items in list appear in order in text.
-
-    Examples:
-      verify_ordered_items_in_text(['a', '1'], 'ab1')  # passes
-      verify_ordered_items_in_text(['1', 'a'], 'ab1')  # raises AssertionError
-    """
-    index = 0
-    for item in to_verify:
-        index = text[index:].find(item)
-        assert index > -1, "Expected item not found: '{}'".format(item)
