+++ tests/integration_tests/modules/test_power_state_change.py
@@ -9,7 +9,7 @@ import pytest
 
 from tests.integration_tests.clouds import IntegrationCloud
 from tests.integration_tests.instances import IntegrationInstance
-from tests.integration_tests.log_utils import verify_ordered_items_in_text
+from tests.integration_tests.utils import verify_ordered_items_in_text
 
 USER_DATA = """\
 #cloud-config
