+++ tox.ini
@@ -153,7 +153,7 @@ deps =
 
 [testenv:integration-tests-ci]
 commands = {envpython} -m pytest --log-cli-level=INFO {posargs:tests/integration_tests}
-passenv = CLOUD_INIT_* SSH_AUTH_SOCK OS_*
+passenv = CLOUD_INIT_* SSH_AUTH_SOCK OS_* TRAVIS
 deps =
     -r{toxinidir}/integration-requirements.txt
 setenv =
